Rearrange data layout
We re-arrange the data layout to "structure to pointers".  However,
there isn't really a structure hold all the pointers.  Instead, we
simply compute the offset.  We also define the macro N_VAR be the
number of variables in the problem.  The bandwidth performance is much
better now (8x on my laptop).