Use shared memory to store G-Pens
The bad performance is probably because we run out of register.  The
gpen[] array was then allocated in the local memory, which is many
times slower than register or shared memory.

To overcome this, we put gpen[] int the shared memory.  I.e., we
explicitly declare

  __shared__ R gpen[N_VAR][1 + 2 * RADIUS][TILE_Y][TILE_X];

Because there are only 16 KiB of shared memory on my laptop GPU, I
have to reduce the number of threads inside the block.

The kernel is at least 2 times faster than the previous version.